+++ fspt27-mvp/client/src/features/gamescreen/rescue/RescuePokemon.jsx
@@ -1,25 +1,25 @@
 import { useState } from "react";
 import { useNavigate } from 'react-router-dom';
-import Catch from "./Catch";
+import RescuePopup from "./RescuePopup";
 import axios from "axios";
 import Modal from "../../../components/Modal";
 import RandomPokemon from "../../../components/RandomPokemon"
 import Menu from "../../../components/Menu";
-import "./CatchLanding.css";
+import "./RescuePokemon.css";
 
-function CatchLanding() {
-  const [isCatchOpen, setCatchopen] = useState(false);
+function RescuePokemon() {
+  const [isRescueOpen, setRescueopen] = useState(false);
   const [randomPokemon, setRandomPokemon] = useState(0);
   const navigate = useNavigate();
 
-  const openCatchModal = () => setCatchopen(true);
-  const closeCatchModal = () => setCatchopen(false);
+  const openRescueModal = () => setRescueopen(true);
+  const closeRescueModal = () => setRescueopen(false);
 
   const handleRandomPokemon = (pokemon) => {setRandomPokemon(pokemon);}
 
   async function updateStorage(pokemon_id) {
     try {
-      const response = await axios.put(`/api/cozygarden/${pokemon_id}?action=increase`);
+      const response = await axios.put(`/api/pocketrescue/${pokemon_id}?action=increase`);
       if (response.status === 200) {
         const data = response.data;
         console.log(`updateStorage() | We have received the request from toStorage and increased this Pokémon:`, data[0])
@@ -47,23 +47,22 @@ function CatchLanding() {
   
     return (
       <>
-        <div id='catchlanding'>
-            <div id='catchlanding-header'>
+        <div id='rescuepokemon'>
+            <div id='rescuepokemon-header'>
               <h2>Rescue a wild pokemon</h2>
               <button onClick={() => navigate(-1)}>x</button>
             </div>
-            <div id='catchlanding-wrap'>
+            <div id='rescuepokemon-wrap'>
               <div id='intro'>
                 <img src="https://64.media.tumblr.com/59f1ea5d708eb9922424f115e97c873b/tumblr_nt1401NDiU1uzvqoio1_r1_540.gif" />
                 <p>Oh no! A wild pokemon was spotted in the forest nearby. It looks really upset. Try and catch it so it can find a warm, forever home.</p>
               </div>
-              <button onClick={openCatchModal}>Catch Pokemon</button>
-              <button onClick={() => navigate(-1)}>Back</button>
+              <button onClick={openRescueModal}>Rescue Pokemon</button>
           </div>
         </div>
-            <Modal isOpen={isCatchOpen} onClose={closeCatchModal}>
+            <Modal isOpen={isRescueOpen} onClose={closeRescueModal}>
               <RandomPokemon onRandomPokemon={handleRandomPokemon} />
-              {randomPokemon && <Catch randomPokemon={randomPokemon} toStorage={toStorage} />}
+              {randomPokemon && <RescuePopup randomPokemon={randomPokemon} toStorage={toStorage} />}
             </Modal> 
         <Menu />     
       </> 
@@ -71,4 +70,4 @@ function CatchLanding() {
   }
   
 
-export default CatchLanding;
+export default RescuePokemon;