+++ fspt27-mvp/client/src/features/gamescreen/rescue/RescuePopup.jsx
@@ -1,10 +1,10 @@
 import { useEffect } from "react";
 import { Link } from 'react-router-dom';
-import "./Catch.css";
+import "./RescuePopup.css";
 
-function Catch({randomPokemon, toStorage}) {
+function RescuePopup({randomPokemon, toStorage}) {
 
-  const nextCatch = () => {
+  const nextRescue = () => {
     window.location.reload();
   }
 
@@ -14,21 +14,19 @@ function Catch({randomPokemon, toStorage}) {
   }, [])
   
     return (
-    <div id='catch-popup'>
+    <div id='rescue-popup'>
       <h2>You have caught {randomPokemon?.name}!</h2>
-      <div id='catch-intro'>
+      <div id='rescue-intro'>
         <img src={randomPokemon?.img} />
         <p>It has been sent to the adoption unit. Better take good care of it so it can quickly find a new home.</p>
       </div>
-      <div id='catch-buttons'>
-        <button onClick={nextCatch}>Rescue another</button>
-        <Link to="/storage"><button>Visit adoption center</button></Link>
+      <div id='rescue-buttons'>
+        <button onClick={nextRescue}>Next rescue mission</button>
+        <Link to="/rescuecenter"><button>Visit adoption center</button></Link>
       </div>
-      <p>{randomPokemon?.caught <= 1 ? `p.s. Congrats, you have rescued your first ${randomPokemon?.name}!` : `` }</p>
-
     </div>
     );
   }
   
 
-export default Catch;
+export default RescuePopup;