+++ fspt27-mvp/model/importPokemon.js
@@ -11,7 +11,7 @@ const con = mysql.createConnection({
   host: DB_HOST || "127.0.0.1",
   user: DB_USER || "root",
   password: DB_PASS,
-  database: DB_NAME || "cozygarden",
+  database: DB_NAME || "pocketrescue",
   multipleStatements: true
 });
 
@@ -19,12 +19,11 @@ con.connect(async function(err) {
   if (err) throw err;
   console.log("Connected!");
 
+  //If you need to import new Pokemon, run 'npm run import-pokemon' to import api data into db
   try {
-    // Haal de basisinformatie op
     const response = await axios.get('https://pokeapi.co/api/v2/pokemon?limit=1500');
     const pokemon = response.data.results;
 
-    // Haal de gedetailleerde informatie voor elke Pokémon op en filter op grasstype
     const detailedPokemon = await Promise.all(pokemon.map(async (p) => {
       const pokemonData = await axios.get(p.url);
       const types = pokemonData.data.types.map(type => type.type.name);
@@ -38,14 +37,11 @@ con.connect(async function(err) {
     }));
     console.log(detailedPokemon)
 
-    // Filter de undefined waarden die niet grasstype zijn
     const grassPokemon = detailedPokemon.filter(p => p !== undefined);
 
-    // Bereid de SQL query en de waarden voor
-    const sqlInsert = 'INSERT INTO pokemon (pokemon_id, name, img) VALUES ? ON DUPLICATE KEY UPDATE img = VALUES(img)';
+    const sqlInsert = 'INSERT INTO pokemon (pokemon_id, name, img) VALUES ?';
     const values = grassPokemon.map(p => [p.id, p.name, p.img]);
 
-    // Voer de query uit
     con.query(sqlInsert, [values], function(err, result) {
       if (err) throw err;
       console.log('Grass-type Pokémon data has been successfully imported');